Extract HardSwish SE layout optimizer
